+++ packages/form-utils/src/index.ts
@@ -1,4 +1,6 @@
 export * from './lib/FieldHelpIcon';
 export * from './lib/FormControlLabel';
+export * from './lib/FormHelperText';
 export * from './lib/FormLabel';
 export * from './lib/Input';
+export * from './lib/Select';