+++ packages/tabs/src/lib/TabList.tsx
@@ -3,21 +3,14 @@ import { TabList as MuiTabList, TabListProps as MuiTabListProps } from '@mui/lab
 export interface TabListProps
   extends Omit<
     MuiTabListProps,
-    | 'centered'
-    | 'centerRipple'
-    | 'disableRipple'
-    | 'disableTouchRipple'
-    | 'focusRipple'
-    | 'orientation'
-    | 'TouchRippleProps'
-    | 'touchRippleRef'
+    'centered' | 'centerRipple' | 'focusRipple' | 'orientation' | 'TouchRippleProps' | 'touchRippleRef'
   > {
   children?: React.ReactNode;
 }
 
 export const TabList = ({ children, ...rest }: TabListProps): JSX.Element => {
   return (
-    <MuiTabList {...rest} disableRipple disableTouchRipple orientation="horizontal" centered={false}>
+    <MuiTabList {...rest} orientation="horizontal" centered={false}>
       {children}
     </MuiTabList>
   );