+++ packages/textfield/src/lib/TextField.stories.tsx
@@ -1,7 +1,8 @@
 // Each exported component in the package should have its own stories file
 import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { EyeIcon, EyeSlashIcon } from '@availity/mui-icon';
+import { Box, Chip, InputAdornment, MenuItem, SelectChangeEvent, Stack } from '@mui/material';
+import { EyeIcon, EyeSlashIcon, SearchIcon } from '@availity/mui-icon';
 import { IconButton } from '@availity/mui-button';
 
 import { TextField, TextFieldProps } from './TextField';
@@ -10,6 +11,14 @@ const meta: Meta<typeof TextField> = {
   title: 'Components/TextField/TextField',
   component: TextField,
   tags: ['autodocs'],
+  args: {
+    helperText: 'Helper Text',
+  },
+  argTypes: {
+    helperText: {
+      type: 'string',
+    },
+  },
 };
 
 export default meta;
@@ -19,27 +28,130 @@ export const _TextField: StoryObj<typeof TextField> = {
   args: {
     label: 'Field Label',
     id: 'test',
+    helpTopicId: '123',
   },
 };
 
+export const _States: StoryObj<typeof TextField> = {
+  render: (args: TextFieldProps) => (
+    <Stack direction="row" spacing={1}>
+      <TextField label="Default" id="default" {...args} />
+      <TextField label="Focused" id="Focused" focused {...args} />
+      <TextField label="Error" id="error" error {...args} />
+      <TextField label="Disabled" id="disabled" disabled {...args} />
+    </Stack>
+  ),
+  args: {
+    margin: 'normal',
+  },
+};
+
+export const _WithIcon: StoryObj<typeof TextField> = {
+  render: () => (
+    <TextField
+      label="With Icon"
+      id="search"
+      InputProps={{
+        startAdornment: (
+          <InputAdornment position="start">
+            <SearchIcon />
+          </InputAdornment>
+        ),
+      }}
+    />
+  ),
+};
+
 export const _PasswordField: StoryObj<typeof TextField> = {
-  render: (args: TextFieldProps) => {
+  render: () => {
     const [showPassword, setShowPassword] = useState(false);
 
     return (
       <TextField
-        {...args}
         type={showPassword ? 'text' : 'password'}
+        label="Password"
         id="password"
-        endAdornment={
-          <IconButton title="password visibility toggle" onClick={() => setShowPassword((prev) => !prev)}>
-            {showPassword ? <EyeIcon /> : <EyeSlashIcon />}
-          </IconButton>
-        }
+        InputProps={{
+          endAdornment: (
+            <InputAdornment position="end">
+              <IconButton
+                title="password visibility"
+                onClick={() => setShowPassword((prev) => !prev)}
+                edge="end"
+                size="small"
+              >
+                {showPassword ? <EyeIcon /> : <EyeSlashIcon />}
+              </IconButton>
+            </InputAdornment>
+          ),
+        }}
       />
     );
   },
+};
+
+export const _Select: StoryObj<typeof TextField> = {
+  render: (args: TextFieldProps) => {
+    const [count, setCount] = useState('');
+
+    const handleChange = (event: SelectChangeEvent) => {
+      setCount(event.target.value as string);
+    };
+
+    return (
+      <TextField value={count} select SelectProps={{ onChange: handleChange }} {...args}>
+        <MenuItem value={10}>10</MenuItem>
+        <MenuItem value={20}>20</MenuItem>
+        <MenuItem value={30}>30</MenuItem>
+      </TextField>
+    );
+  },
   args: {
-    label: 'Password',
+    label: 'Select',
+  },
+};
+
+export const _MultiSelect: StoryObj<typeof TextField> = {
+  render: (args: TextFieldProps) => {
+    const [multiValue, setMultiValue] = useState<string[]>([]);
+
+    const handleChange = (event: SelectChangeEvent<typeof multiValue>) => {
+      const {
+        target: { value },
+      } = event;
+      setMultiValue(
+        // On autofill we get a stringified value.
+        typeof value === 'string' ? value.split(',') : value
+      );
+    };
+
+    const options = ['one', 'two', 'three', 'four', 'five', 'six'];
+
+    const renderValue = (selected: typeof multiValue) => (
+      <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.5 }}>
+        {selected.map((value) => (
+          <Chip key={value} label={value} />
+        ))}
+      </Box>
+    );
+
+    return (
+      <TextField
+        id="multiple-chip"
+        select
+        SelectProps={{ multiple: true, onChange: handleChange, renderValue: renderValue }}
+        value={multiValue}
+        {...args}
+      >
+        {options.map((option) => (
+          <MenuItem key={option} value={option}>
+            {option}
+          </MenuItem>
+        ))}
+      </TextField>
+    );
+  },
+  args: {
+    label: 'MultiSelect',
   },
 };