+++ packages/form-utils/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.4.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.3.0...@availity/mui-form-utils@0.4.0) (2023-09-26)
-
+## [0.5.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.4.1...@availity/mui-form-utils@0.5.0) (2023-10-20)
 
 ### Features
 
-* **mui-button:** icon button styling ([43c424e](https://github.com/Availity/element/commit/43c424e2f33b0dcd9dca51ec253e54ea3309630b))
+- **mui-form-utils:** add FormHelperText and Select ([0ff9d8e](https://github.com/Availity/element/commit/0ff9d8e06d0baac263ce1869304d3b91fcb5cb15))
+
+## [0.4.1](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.4.0...@availity/mui-form-utils@0.4.1) (2023-10-06)
+
+## [0.4.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.3.0...@availity/mui-form-utils@0.4.0) (2023-09-26)
+
+### Features
+
+- **mui-button:** icon button styling ([43c424e](https://github.com/Availity/element/commit/43c424e2f33b0dcd9dca51ec253e54ea3309630b))
 
 ## [0.3.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.2.0...@availity/mui-form-utils@0.3.0) (2023-09-25)