+++ packages/icon/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.6.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.5.0...@availity/mui-icon@0.6.0) (2023-09-26)
-
+## [0.7.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.6.1...@availity/mui-icon@0.7.0) (2023-10-20)
 
 ### Features
 
-* **mui-icon:** add xsmall fontsize ([aa8e28a](https://github.com/Availity/element/commit/aa8e28a5a7514247db0f65d9120da2847663445a))
+- **mui-icon:** add expand/collapse icons ([0a88920](https://github.com/Availity/element/commit/0a889203e15785beca2834d79d6d2bfc3425bfcd))
+
+## [0.6.1](https://github.com/Availity/element/compare/@availity/mui-icon@0.6.0...@availity/mui-icon@0.6.1) (2023-10-06)
+
+## [0.6.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.5.0...@availity/mui-icon@0.6.0) (2023-09-26)
+
+### Features
+
+- **mui-icon:** add xsmall fontsize ([aa8e28a](https://github.com/Availity/element/commit/aa8e28a5a7514247db0f65d9120da2847663445a))
 
 ## [0.5.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.4.1...@availity/mui-icon@0.5.0) (2023-09-25)